minio_upload/requirements.txt
numpy
python-dotenv
minio
omegaconf